SDC-6524. Avro Parser Cannot read logical types properly.
This patch fixes the case where the avro parser sees value for a logical decimal type in bytes and cannot handle this as DecimalTypeSupport can only handle numbers/strings, given that this is specific to avro, we will use the reverse process from how we wrote as bytes to read it back as BigDecimal.

Also for date we follow for serialization what hive follows convert date to millis and to number of days in integer to epoch and while deserializing it back which convert from number of days to millis ato date, this patch handles that as well.

Change-Id: Id6b76c88efbae321be994248524c9110a77ecd85
Reviewed-on: https://review.streamsets.net/8954
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
(cherry picked from commit 93314e7f5a602c2016d95b86e2ae0b13eaec4f92)
Reviewed-on: https://review.streamsets.net/8968
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>