SDC-6540.  Field hasher returns incorrect value.
Added a new checkbox to the UI "Use Field Separator"
to toggle between old and new behavior.

When the checkbox is enabled, the old behavior will add
a null as a separator and a field terminator. This provides more
uniqueness (see DeDuper comment below) but gives unexpected results
when customers try to duplicate the hash value with a stand-alone
Java program or some other mechanism.

With the checkbox off, no nulls are added; the fields are
simply concatenated then hashed, this provides the
expected hash value and the hash values will
match what the customer calculates with a Java program.

When creating a new pipeline this option defaults to false
so FieldHasher will operate "new way".  Upgrading an
existing pipeline will result in this option being set to true
which will enable "old way" (with field separators).

DeDup processor depended on this hashing code too, and DeDup
has been enhanced to work strictly old-way. Since the
hash values are not persisted, this is not a problem.
Actually, this is a feature, since placing the nulls can
provide more "uniqueness" to the hash string.

For example, without the null as a separator,
field1 with "abc" and field2 with "def"
will hash to the same value as field1 with "ab"
and field2 with "cdef", possibly giving a false positive.
The null acts as a field separator for DeDuper.

Change-Id: I1d8bdd019a376d35b2fa4f54a688a562b6de2ce3
Reviewed-on: https://review.streamsets.net/8997
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Hari Shreedharan <hshreedharan@streamsets.com>