SDC-6637: Convert SupportBundleManager to Task
This enables us to decouple object creation and initialization - thus
the support bundle manager better plugs to the start up routine. One
immediate effect is that logs from the initialization gets properly
routed to the sdc.log - they weren't before as the object was created
before we initialized logging infra.

Change-Id: I8074d939a0a7d7c9e904cdd7d2f4c6d997fdb1d5
Reviewed-on: https://review.streamsets.net/9080
Reviewed-by: Bob Plotts <bob@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>