SDC-6393: Make error record behavior configurable
This patch makes the behavior of error records configurable - by default
we will behave the same way as before, that is the record persisted in
error stream will be the same as it was generated by origin.

This patch also adds a second policy - persist the record as-it-was when
it was sent to error stream (e.g. after all the processing that happened
inside the pipeline).

Implementation wise, I've moved some logic from ProductionPipelineRunner
into the BadRecordsHandler as that seemed abetter way to deal with error
records rather then having all-mighty runner that deals with everything.

Change-Id: I0da3038847a24280553a7e39cd35a4f60f80a6fa
Reviewed-on: https://review.streamsets.net/9081
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>