SDC-6649: Lineage: Propagate events to the actual publisher in LineagePublisherTaskImpl
This patch concludes the main set of changes for the publisher task -
it's now capable of buffering events in in-memory queue and consuming
them in a separate thread by sending them to the lineage publisher.

This patch currently doesn't validate if the events were properly
consumed by the publisher (e.g. checking return value of the publisher's
publishEvents method). That will be added in subsequent JIRA.

Change-Id: If12e2e943212ff7f87cf60f6f796311f61ddb0d8
Reviewed-on: https://review.streamsets.net/9098
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Bob Plotts <bob@streamsets.com>