SDC-6655: Provide test specific builder for Pipeline.Builder
I had two occasions where I need to add a new parameter to the
constructor and it required me to go over ~20 test files and change them
by adding a mocked value. Also as the constructor grows, it's hard to
see what is a "dump" value and what actually matters for the test, so I
would like to propose to create a test specific builder that will fill
the default values that are applicable for tests. Then each test case
can override the ones relevant for that particular test which will make
it obvious to the reader.

Change-Id: If248bc85495ec74780d4dc1925ff722fc774008b
Reviewed-on: https://review.streamsets.net/9101
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Bob Plotts <bob@streamsets.com>