SDC-6176. Kafka Consumer With Confluent Schema Registry uses the old schema until pipeline restart
Issue was that we assume, if we have a schema, that it is the correct one.
Saving the schema ID and comparing it to the detected one allows us to load
a new schema at the appropriate time.

Change-Id: Ib0c5a8db5075ec1daca4ef4668e175daa7411463
Reviewed-on: https://review.streamsets.net/9113
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Adam Kunicki <adam@streamsets.com>
(cherry picked from commit b6b0cf312066a1303e168214284bdf4a05c7504a)
Reviewed-on: https://review.streamsets.net/9127
Reviewed-by: Pat Patterson <pat@streamsets.com>