SDC-6468: Return any applicable properties to Local FS
The root of the trouble is that the fileType configuration that the
missing properties are depending on is hidden in the Local FS
destination.

I believe that we've hidden the fileType originally because it had two
values - Text and SequenceFile - and Sequence file is not much
applicable for local fs. However today the list also contains the Whole
file support.

Hence rather then rolling back the SDC-6169 I think that better solution
is to un-hide this config and as a nice side effect allow the local fs
to work with whole files.

Change-Id: I6f22f6e1c571bcbff4adf6ef3f1880842c8e6a3d
Reviewed-on: https://review.streamsets.net/9128
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Junko Urata <junko@streamsets.com>