SDC-4385: Support XML in Destinations
This patch connects already existing logic that serializes record to XML
with our DataGeneratorFactory logic and configs so that it can be
plugged into a stage. It also adds the XML format to Kafka stage with
proper tests to make sure that it works as expected.

There are few follow ups that we might or might not resolve as future
tasks:

* Since our XML reader is very optimistic and parsed XML looks
  differently then the one generated, writing -> reading does not
  produce the same input as as written.
* Generated XML will be invalid/unparsable if user selects the "One
  Message per Batch" config option.

Change-Id: Ic06465b09228ca73ff15ab7989cded5f73c23219
Reviewed-on: https://review.streamsets.net/9139
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>