SDC-5174: JSON data format with extra space from the second line
So the trouble is that when serializing records in MULTIPLE_OBJECTS
mode, were were actually inserting the new line characters ourselves and
kept the JSON writer in the default setting. But in default setting it
will use space a record separator and hence we ended up with two
separators - new line that we inserted ourselves and space that JSON writer
inserted.

 This patch changes that behavior by configuring the writer to use new
 line as a separator and removing the code on our side to write the new
 line character. This way there is only one entity responsible for
 writing the record separators and it's the correct one.

 The patch also adds tests as there were sadly missing for this class.

Change-Id: I16c63066ab166f8b11c862900224a7e18e8b6a3c
Reviewed-on: https://review.streamsets.net/9234
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Junko Urata <junko@streamsets.com>