SDC-5228: Preview with Write to Destination enabled doesn't work with Text Data Format in destination
This patch adds a new select box for Text file format - action on what
to do when a field is missing. Right now the implicit behavior is
"ignore" which is very confusing to users as it's pretty much
un-intentional data loss. This existing option is kept, but a new option
is provided that is also a default value - throw an error. This way records
that do not have configured field be propagated up to stage and dealt
with accordingly - mostly likely sent to error stream.

This patch sets the default value to error regardless whether this is a
new pipeline or upgraded one which means that it's technically a
breaking change. Given the fact how many users were surprised by the
behavior, I was thinking about doing this intentionally to let SDC
scream at people in case that their pipeline is mis-configured (e.g.
it's intended as a bug fix rather then breaking previous "normal"
behavior).

Change-Id: I8237c3b92332b66b504a7dd4445940770876bff8
Reviewed-on: https://review.streamsets.net/9239
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>