SDC-6449. java.lang.LinkageError: com/codahale/metrics/Counter
The logic for this fix is
- Get the version of metrics-core from sdc_home/api-lib.
 (It is "3" in the current SDC)
- Look for a symlink to metrics-core and remove it if version is same.
I didn't consider sub-version here, so for example 3.0.0 and 3.10.10
are considered to be same version and remove symlink.
If there is extra characters after or before the version number
(i.e metrics-core-3.0.1-mapr-blah.jar), I treated it as
different jar so keep the symlink.

Change-Id: Iaee814f0763f43eef07c124ab6623164728da9fd
Reviewed-on: https://review.streamsets.net/9241
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Junko Urata <junko@streamsets.com>