SDC-6780: Pipeline Lifecycle: Add new stage lists to PipelineConfiguration
This patch adds the startEventStages and stopEventStages to the
PipelineConfiguration object that represents pipeline the pipeline
structure (on disk and over REST). Sadly the class is used directly in
quite a few files (mostly tests), so I've decided to submit this patch
separately. Proper Pipeline upgrade will be done in a follow up JIRA.

Change-Id: I821cd8eb5305db19bd11169e093404a2a9c3a9cc
Reviewed-on: https://review.streamsets.net/9365
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>