SDC-6810: Further improve SdcInfoGenerator to selectively ignore errors
I got ahold of a stack trace suggesting the root cause of the corrupted
support bundle. It does suggest a broken symlink. We've partially
improved this by SDC-6785 by not terminating the whole generator, but
rather an individual file listing. I would like to take this one step
further and just catch the given metadata call, and place ERROR on its
place.

Change-Id: I813d8107da1aeea3388dc720e2c0350903c85304
Reviewed-on: https://review.streamsets.net/9412
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Ji Sun Kim <jisun@streamsets.com>