SDC-6836: Pipeline Lifecycle: Display pipeline lifecycle event configuration in UI
This is my stab to at least somehow display the configuration for
start/stop events in the UI. I'm not a JavaScript guru, so feel free to
comment and suggest if I should take different path. Heck, I'm happy to
discard this if someone more JS-friendly want to write it differently.
I'm sure that we tweak this as we go anyway.

I've pretty much copy-pasted logic that we had for error/stats
aggregation, tweaked it slightly for lifecycle events and tested that it
works as expected.

Change-Id: I03a15dc561245eabe407ad3e6a5e404c74474394
Reviewed-on: https://review.streamsets.net/9462
Reviewed-by: Madhukar Devaraju <madhu@streamsets.com>