SDC-6848: Pipeline Lifecycle: Validation errors displays on both start/stop handlers if using same executor
I've noticed that when I used the same executor for both start and stop
event handling and made a configuration issue in only one of them, the
error was also displayed in the other one. Please see JIRA for
screenshot.

Looking into the code and following the logic, I think that we missed an
if condition that was checking stage instance name.

Change-Id: I00170d2a89608a30c7fe8980d567fff84fd4534e
Reviewed-on: https://review.streamsets.net/9468
Reviewed-by: Madhukar Devaraju <madhu@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>