SDC-6849: Pipeline Lifecycle: Validate that ToError and it's stage have the same version
After many upgrade issues, we've instrumented the code to have a comment
to notify the developer when upgrading stage that also has subclasses
that it needs to upgrade their version as well. It seems that the
comment is not very effective as people are not following it anyway.
Hence I would suggest adding a test code that will ensure that - this
way the pre-commit hook will fail and presumably the engineering will
fix that before merging.

Change-Id: I99cd768441bc4c11ba15e09ecc14ada648ca60d3
Reviewed-on: https://review.streamsets.net/9469
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Adam Kunicki <adam@streamsets.com>