SDC-6842: SDC wont start and throws RuntimeException with 'Missing CredentialStore definition 'streamsets-datacollector-jks-credentialstore'
Core of the trouble is that we added a default configuration that
assumes that certain stage library is always present. That however does
not directly corresponds with our installation paths as both STF docker
image and core tarbal does not ship with this new stage library.

Looking at the default configuration, it does expect a JKS file to be
available that user actually needs to create in order to make this
functionality helpful. Hence I'm thinking that the best solution is to
simply disable the default store by default and put instructions how to
enable it (after user also creates the actual keystore).

Change-Id: Ied55be594a9c8bfcbb2b65c0ff99c1e64ec48ca4
Reviewed-on: https://review.streamsets.net/9485
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Hari Shreedharan <hshreedharan@streamsets.com>