SDC-6860: Reading Avro Data type - automatically infer type for null union
There is a fundamental difference between Avro null and SDC Record null:

* In Avro null is a type - e.g. it's on the same level as string or long
* In SDC (and in Java in general) null is not a type, it's value instead -
  e.g. String object can have value "Hi" or have a value null.

We're converting Avro null to map as we have to map it to something -
and that something needs to be a constant. So map is an as good
candidate as anything else.

There is, however, one improvement that this patch introduced - since Avro
type ["null", $type] is extremely common Avro idiom to specify that certain
field has given type or null - we simply create typed null.

Change-Id: Icbd10e8301521c6615e4f31767ef7e73f214286b
Reviewed-on: https://review.streamsets.net/9496
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Ji Sun Kim <jisun@streamsets.com>