SDC-6800: Pipeline Lifecycle: Properly handle lifecycle events in preview mode
A bit larger change because it touches both UI and backend - but it's
straightforward in it's nature. We've added a new property in the UI
that enables user to selectively enable or disable the lifecycle events
when doing preview.

This way user can limit side effects that might or might not be
important for the pipeline. I was originally thinking about using the
existing check box we have for "writing to destinations", but it does
have a slightly different semantic - e.g. user might need to run
lifecycle events because they alter the origin's state, but might not
want to write to destinations.

Change-Id: I1b9ef6c54f8a904f5b2a2780dd14fc8e4fbc8a28
Reviewed-on: https://review.streamsets.net/9499
Reviewed-by: Jeff Evans <jeff@streamsets.com>