SDC-6867: Pipeline Lifecycle: Stages producing events throws NullPointerException
The trouble is that we were not passing proper Error and Event sinks and
hence the NPE when we tried to use them. We have validation that the
stage does not have connected event lane, so ignoring the generated
events is correct thing to do. For error sink, we might need to do
follow up work of detecting if the event failed processing and throw
exception accordingly - but that will be done as a separate work as it
might be influenced by decision that we need to do on SDC-6844.

Change-Id: Ie1d6b0af159f52984aa164c3cb7d549db04b30aa
Reviewed-on: https://review.streamsets.net/9507
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>