SDC-6813: Support new logical types introduced in Avro 1.8
This patch added support for all new time-based logical types that were
introduced in Avro 1.8. Since they are logical types, we actually don't
need updated Avro libraries and hence the whole change is a bit easier.

Avro 1.8 also added a logical type called "duration" - it's a weird type
and I'm not entirely sure what is the best way to represent it. Let's
wait on an use case that will need it before adding it.

Change-Id: Id1117c27416bdfca2e0c715a528ba8d666b6f272
Reviewed-on: https://review.streamsets.net/9553
Reviewed-by: Junko Urata <junko@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>