SDC-6889: Fix license header round 2
I've followed the same procedure as last time in SDC-6435. Just instead
of running the conversion on whole project, I run it only on modules
that had incorrect license headers. This way UI and other stuff should
not broke like last time

For posterity, here is the procedure:

  mvn com.mycila:license-maven-plugin:3.0:format  -Dlicense.header=license_header.ftl

The content of license_header.ftl is as follows:

Copyright 2017 StreamSets Inc.

  Licensed under the Apache License, Version 2.0 (the "License");
  you may not use this file except in compliance with the License.
  You may obtain a copy of the License at

      http://www.apache.org/licenses/LICENSE-2.0

  Unless required by applicable law or agreed to in writing, software
  distributed under the License is distributed on an "AS IS" BASIS,
  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
  implied.
  See the License for the specific language governing permissions and
  limitations under the License.

Change-Id: I9f67db23f84e3f145df74ed7f3a294a01e5af391
Reviewed-on: https://review.streamsets.net/9571
Reviewed-by: Anthony McKeehan <anthony@streamsets.com>