SDC-6915: Synchronize alteration queries in Hive Metastore Target
While testing JDBC Scale up with Hive solution, I noticed that more
often than not, my pipelines failed with Hive Metastore Target throwing
an exception that "Table already exists."

This is because all the different runners were ingesting the same table
and as there is no synchronization, the different instances of the
target tried to create the table at the same time.

This patch introduced locking between various runners to make sure that
only one of them will be working with given table at a time. E.g. runner
2 will wait on runner 1 if both wants to work with table "x", but if one
is working on "x" the other on "y", then they will happily work in
parallel.

Change-Id: I6b280f24cd9f994d1b1a37b8a678f9ca15eea2c9
Reviewed-on: https://review.streamsets.net/9573
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>