SDC-6758: Pipeline with kafka origin fails if produce single record is on and record is send to error handler
I've found similar code fragment 3 times in the code base, so I've
changed all instances to look roughly the same. I have not abstracted it
to a function as there is no really good place where to put the shared
code - seems reasonable as the code fragment is not that large.

Change-Id: I65b82e29ca4bd884d6cadc2a121dd3dd73d04711
Reviewed-on: https://review.streamsets.net/9591
Reviewed-by: Virag Kothari <virag@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>