SDC-6914: TestDataParserProcessor.testSyslogParsing constantly fails
The issue with the failing test lies in parsing of RFC3164 date. The
RFC3164 is a bit weird - it contains date and month, but no year. So we
have to guess the year somehow - and that guessing logic was the one
that was failing.

The previous code was based on really old version of Flume - and even
Flume project improved the logic via FLUME-2889 some time back. The
current implementation follows the logic that Flume introduced - instead
of hardcoding +/- month of valid dates, it creates a sliding window that
goes back 11 months and 1 month to the future. Messages from this window
will have correct year. Message outside will have incorrect year - but
since the message itself doesn't contain year, there is not much that we
can do better.

Change-Id: I8eee61424b756a1d6c8b3b81376cd60d332d5879
Reviewed-on: https://review.streamsets.net/9598
Reviewed-by: Bob Plotts <bob@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>