SDC-6930: TestDataParserProcessor.testSyslogParsing throws NPE
This is a follow up from SDC-6914. The test now can throw
NullPointerException. The NPE is actually coming from test code that
assumes that certain structures are there, the real issue is that the
date is not correctly parsed.

And that is actually caused by incorrect test data as it's already
August 1 in UTC timezone – and the generated date is Aug 1 when based on
the RFC it should be Aug 1 (two spaces instead of one).

Change-Id: I6f2897649f1923f11701444a9c50a143be81c55e
Reviewed-on: https://review.streamsets.net/9608
Reviewed-by: Jeff Evans <jeff@streamsets.com>