SDC-6429: JDBC Query origin never issues no-more-data if queryInterval is 0
I've tried to simplify the logic a bit while also making sure that the
no-more-data event will be issued even when there is no delay
configured. The idea is that if we run a query this batch and the query
returned empty, then we reach point where we need to trigger no-more-data
event.

The logic in the query origin is slowly getting quite complicated, so
let me know if I missed some nuanced behavior while making this change.

Change-Id: I12d8da930f3f3ab7134ca7ca2e3aee20017ce38f
Reviewed-on: https://review.streamsets.net/9623
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Bob Plotts <bob@streamsets.com>