SDC-6919: Start up sometimes fails with NoSuchMethodError:com.google.common.collect.Sets$SetView.iterator
I'm not sure why it "sometimes" works and sometimes does not but the
core of the problem is that our swagger depends on package called
reflections on version 0.9.11. This version however depends on guava
0.20 while container ships only 0.18.

And the reason why we started seeing this only recently is that we've
just recently upgraded swagger - SDC-6835..

We have two options - either upgrade guava or downgrade back the swagger
dependency. I've tried to upgrade guava, but with current build it's
very tricky. So to not block 2.7 release, this patch reverts swagger
back to the last version that still depends on reflections 0.9.10.

Change-Id: Ifb0d2fa5b92a40deb3f1c28c4ba4567b57a17949
Reviewed-on: https://review.streamsets.net/9579
Reviewed-by: Jeff Evans <jeff@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>