SDC-7001: The SDC doesn't automatically generate the heap dump if there is an older heap dump file in the directory
When I introduced the automatic heap dump creation back in SDC-4851, I
have not realized that JVM will not override older dump if it already
exists. Which makes this feature a bit limited in it's use.

This patch slightly alters the default value to include number of
seconds on SDC start. This way each start will dump a different heap
dump if needed.

Change-Id: Id3f99b387ea79073b14e5ac18631e120b2ba9019
Reviewed-on: https://review.streamsets.net/9692
Reviewed-by: Bob Plotts <bob@streamsets.com>