SDC-7065: Use field attributes for decimal metadata
While working on SDC-3954, I came to a need to get access to decimal's
metadata - precision and scale - that sadly needs to be passed out of
the band (not with the data themselves). We solved this problem already
in JDBC components by using record attributes, but that doesn't work
well for nested types. Looking around Avro parser already stores the
decimal and precision in field attributes (instead of record attributes)
which is really a good place where to store such metadata.

Change-Id: I573f26d70882f16f90c134f52d36ad7b12f6a1b6
Reviewed-on: https://review.streamsets.net/9819
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jeff Evans <jeff@streamsets.com>