SDC-7080: Support specifying default values per field type / field path in Schema Generator
Added handlers for default values. The patch introduces two ways how to
set a default value:

1) If user enabled "nullable" types, he can also enable setting the
default to null.

2) Added a select box that for each type will set the default value to a
constant. Adding a default value for a type will override any null as
default (which seems like an expected behavior).

This patch does not currently contain a field path based setting of a
default value as that actually seems quite tricky (to escape the patch
100% properly with the current implementation). We can add that later
(hopefully after we a bit formalize what can be used in field path and
what characters can not).

If nothing above then no default value will be used.

Change-Id: Iadf1df229ae32626c84a368dd8d122342f2c9e03
Reviewed-on: https://review.streamsets.net/9909
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jeff Evans <jeff@streamsets.com>