SDC-7217: Pipeline time rules do not work properly with multi-threaded pipelines
Currently the metric rules for time in batch and time in stage are only
using runtime metrics gauge that contains unified info for all pipeline
runners. Hence it could easily happen that one thread is stuck in a
batch for hours while another thread is constantly updating the central
gauge and thus an alert will never be thrown.

This patch removes the central notion of "time in a batch" and "time in
stage" and converts the rule to look into each pipeline runner metrics
instead.

Change-Id: I2073edb219236236575b7bd8a9eb0e7b1b55e826
Reviewed-on: https://review.streamsets.net/10210
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jeff Evans <jeff@streamsets.com>