SDC-5933: Possibility of using Views as source objects in the JDBC Multi Table Origin
This patch enhances the Multi Table JDBC origin to work with views as
well. The primary use case is to only support "simple" views. E.g. views
that selects from a single table and are typically used for column
security (to omit some sensitive columns). Whereas this patch won't die
on complex views (like joins, ...), it's highly not recommended to use
this with complex views as we will run several queries (likely even in
parallel) that might cause too much load on the source database.

Change-Id: Icb697ba03ed5f3e3380b137c0c754d9edbcec10c
Reviewed-on: https://review.streamsets.net/10277
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>