SDC-7175: Google Pub/Sub to Big Table pipeline fails
This patch implements the recommendation [1] from BigTable authors and
overrides all netty jars with a custom version. This custom version is
in sync with what google pub/sub ships and hence everything is working.

This is fairly brittle approach, but I'm not sure if we can do better
given the fact that it's the vendor (Google) who doesn't ship mutually
compatible artifacts without this manual override.

I've tried in one iteration to merge the bigtable and google-cloud
libraries together to have at least everything contained in the same
stage library, but sadly that doesn't work either as the pub/sub and
bigtable have mutual incompatible dependencies (versions of grpc and
google-auth doesn't fit together and sadly different versions does not
play nicely together).

Change-Id: Ia57996d62d8509611c6999bd411c19861b6e5d61
Reviewed-on: https://review.streamsets.net/10291
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>