SDC-7285: CredentialValue: Convert OAuth2 jwtClaims field back to String
This patch converts jwtClaims back to STRING from CREDENTIAL since it
requires explicit evaluation - the TimeEL is indeed required. The field
is a hybrid - it contains some non-sensitive information, but depending
on the OAuth2 provider it can contain tokens. After chatting with our
in-house OAuth2 expert, we've decided to remove the VaultEL from there
for now.

Change-Id: I0be1c949b844332b1eb96604399843fcacb90354
Reviewed-on: https://review.streamsets.net/10314
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Ji Sun Kim <jisun@streamsets.com>