SDC-4463. Native support for systemd
This adds an additional core RPM build.
This also updates the uber rpm tarball script
to rely on symlinks rather than disk intensive copying for
staging the rpms. It also disables gzip compression on the tar
file as rpms are already compressed with cpio.

Ultimately we should stop making an rpm tarball in favor of
publishing a yum repo.

**This change may require updates to Jenkins jobs**

Change-Id: I711646d00b05318291a724dc38433b9457659d87
Reviewed-on: https://review.streamsets.net/10335
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Rupal Shah <rupal@streamsets.com>
Reviewed-by: Ji Sun Kim <jisun@streamsets.com>