SDC-7328: Declare Kafka as provided in kafka protolib
t seems that our proto ib kafka_source-protolib added a dependency on
kafka-clients directly in the protolib which auto-included it in all
child stage libraries. This seemed fair at the time we were doing this,
but it has the unforeseen consequence of being included in MapR stages
where the jar should not be.

Looking around it seems that we generally follow logic to declare the
kafka-protolib dependency as provided in the protolibs and declare it
with given version in the actual stage libraries, so I propose to do the
same here as well.

Change-Id: Ia2f2b9059ec28c3effd64f43ec1bbd785d05326a
Reviewed-on: https://review.streamsets.net/10426
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Bob Plotts <bob@streamsets.com>
(cherry picked from commit a391ea507460b20392cb80faed6c3ed2fc9fe458)
Reviewed-on: https://review.streamsets.net/10498
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>