SDC-6716: Kafka Consumer queue should honor max batch size
Rather then using the fixed constant of 10k, this patch changes the
pre-queue to be as large as it's the batch size itself. The best
possible solution would be to expose this config to the user, but since
the configs are reused in other Kafka origins where this queue is not
present, the user experience would be weird. The new Multi Topic Kafka
origin which is the future where we're going with Kafka origins does not
need pre-reader queue at all, so this is more of "hotfix" rather then
permanent solution.

Change-Id: I8435bf9429d4b16ab8877784ac74f16e4755a0b2
Reviewed-on: https://review.streamsets.net/10429
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Adam Kunicki <adam@streamsets.com>