SDC-7355: Failure Snapshot: Remove CombinerPipe
This patch removes CombinerPatch as it fundamentally does something that
the FullPipeBatch already have to do - combine multiple input pipes into
a single output stream. So rather then doing that explicitly in it's own
pipe, we do that as part of StagePipe processing. This way we don't have
to (shallow) copy all records for each stage.

Change-Id: I8f5c738e63f5a220c56543e750f20c2a8ba3bd72
Reviewed-on: https://review.streamsets.net/10460
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jeff Evans <jeff@streamsets.com>