SDC-7486: Modularize Kafka Multitopic stage library for multiple Kafka versions
This patch modularizes the proto library for multitopic Kafka consumer
and moves the "KafkaConsumer" (Kafka class) into standalone Java loader.
This way child projects can supply version that is compiled against
required Kafka version.

I've decided to go with a new set of maven modules, so that the older
structures can as-they-are be removed in the future when the older
origin will be deprecated and subsequently removed.

Change-Id: Iecfc01bf504f5b2cc227b429761f2c3775216882
Reviewed-on: https://review.streamsets.net/10758
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>