SDC-7500: Service: Add dataformats stage library to core package
This is important because the core tarball already contains basic-lib
that does have bunch of stages with data format tabs that will get
offloaded to the services concept. Hence the dataformats stage library
needs to be shipped as part of the core package, otherwise the package
would be useless (wouldn't start on the account of missing
dependencies).

Change-Id: I74ef94682e939524ad635f1e383f537bf9d22957
Reviewed-on: https://review.streamsets.net/10768
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>