SDC-7501: Service: Create barebone Data Generator procesor
In order to property test how various pieces fits together, I need to
have a processor that actually depends on a service. Rather then
breaking existing stage, I've decided to add a completely new one that I
can co-develop while working on the services. The benefit of this is
that we can easily comment out the @StageDef if we will need to release
from master branch without actually blocking a release on the services
feature.

Change-Id: Ibecbae03d7cdaa752c91c3545ee05638d66de2dc
Reviewed-on: https://review.streamsets.net/10769
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>