SDC-7534: Services: Add services section to main pipeline definition file
This patch changes the on-disk structure of pipeline to contain a new
field called "services" for each stage in the pipeline definition. Since
we've changed the structure, the pipeline schema version was increased
to 5 and proper upgrade have been added as well.

After this change, pipelines generated with this change won't be
importable to version without this change. E.g. the patch is backward
compatible only. Since that is our usual stand (and all previous 4
schema bumps had same deal), it shouldn't cause much troubles outside of
folks playing with "random" nightly builds.

Change-Id: I5731d99768e512b7337f50af9e3740da0d05e329
Reviewed-on: https://review.streamsets.net/10804
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>