SDC-7544: Service: Enable editing service configuration in the UI
For purpose of letting user configure the service, we display all
configuration in the same place - e.g. all service tabs shows next to
the stage normal tabs. This patch assumes that the tab names will be
disjunct which is a validation that will be added in subsequent patch.

Validations won't work yet as the backend itself is not capable of
validating service configuration yet and therefore will also be added in
subsequent patches.

Change-Id: I67e6d432a8f4a2c32defecfd5e41d00b7db46e45
Reviewed-on: https://review.streamsets.net/10819
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>