SDC-7554: Service: StageLibraryTask needs to be able to re-create ServiceDefinition with new private class loader
While working on initialization of the service runtime structures I
realized that the current interface of StageLibraryTask is not on parity
with normal Stage concepts. It doesn't know whether given structure is
requested for an execution and thus might need duplicated class loader
in case that the service definition requires private class loader. This
patch updates the API to be on parity.

Change-Id: Icbd31e59e06a95a93393180ca796e7c56495ebcf
Reviewed-on: https://review.streamsets.net/10830
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>