SDC-7576: Service: Create ConfigInject out of PipelineBeanCreator
While trying to generate runtime structures for services, I realized
that reusing the existing PipelineBeanCreator methods is quite hard
because it's too tailored to the concept of Stage. This patch extracts
the logic for ingesting configuration values and abstracts it in a way
that we can easily re-use it for services.

Change-Id: Ib95585abca4db4eb578fb811c23e43459e5e36ce
Reviewed-on: https://review.streamsets.net/10839
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>