SDC-7579: Service: Add service concept to Issues
This patch enhances our Issue facilities to understand concept of a
service. The implementation follows similar rules as we already have for
pipeline and stage - if the serviceName is null, then the issue is for
the stage itself. If the serviceName is not null, then it's for a
service (of given stage).

Change-Id: Ie850438b9237f591269a0777cf7d0f750030c033
Reviewed-on: https://review.streamsets.net/10850
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>