SDC-7596: Services: Cluster mode pipeline fails with unresolved services
Our Cluster Manager will scan the cluster pipeline and will send to the
backend only those stage libraries that are required/used by the given
pipeline. Hence we need to enhance this code to also look for all used
services and send their appropriate stage libraries as well.

There is one twist to it though - sending service libraries only for
"used" services is not sufficient. The current code will send to backend
the entire stage library even if only one stage out of is used. Best
example is basic-lib - even if only Trash is used, the whole library is
sent with all it's 30+ stages. And since our start up procedure loads
all the available stages (not only those "required") we need to send
service depenencies for all shipped stages, not only for the "used"
ones.

Change-Id: Ic7bc32f1bf8d63a15ddd773814570268959ba982
Reviewed-on: https://review.streamsets.net/10885
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Bob Plotts <bob@streamsets.com>