SDC-7650. Cache KuduRecordConverter to increase performance
KuduRecordConverter object is created per table in every batch. If all records in batch are written
to the same table, we create the same object per batch. The size of this object may become big
if a table has many columns. This fix is to cache KuduRecordConverter so that we don't create/delete
same object often.
This test is based on one pipeline using dev -> kudu with a table with 15 columns.
Record/sec: 750~780  -> 800~810
Heap Usage: size of usage is about the same but the previous code has more activities

Change-Id: I64ac540b33e3e77c41cd037bdaba6bdd90d9a8ca
Reviewed-on: https://review.streamsets.net/10993
Reviewed-by: Junko Urata <junko@streamsets.com>